Refactor CLI command helpers into SRP modules
